pyproject.toml
[project]
name = "qmsr"
description = "{description}"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "absl-py",
  "numpy==2.1.3",
  "jaxlib",
  "jax==0.4.35",
  "scipy",
  "tabview",
  "h5py",
  "jax",
  "setuptools",
  "matplotlib",
  "exp @ git+https://git@github.com/Algopaul/exp",
  "gputil @ git+https://git@github.com/Algopaul/gputil",
]

dynamic = ["version"]


[project.optional-dependencies]
dev = [
  "pylint>=2.6.0",
  "yapf"
]

[tool.setuptools]
packages = ["qmsr"]
